requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==2.3.3
Flask-Cors==4.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-dotenv==1.0.0
requests==2.31.0
setuptools==75.1.0
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.44.0
gunicorn==23.0.0
